Refactored version almost caught up
Buttons have been implemented. Buttons are now capable of having a spritesheet due to implementation of button hover and button click (Button click is very brief so probably best not to have a sprite for it).
Settings is in the process of being implemented.
Spritesheets have been implemented; However, it has not yet been tested. Will be tested when checkboxes are implemented.