Added score system - Will display highest score achieved on Main Menu and current score on the Game screen Added Links to all credits, Bartosz and Connor leads to GitHub, Nathan leads to own website
Changed Direction enum values to relate to their angle in degrees
Changed how entities are rendered, if the entity has no texture property, it will display a rectangle, if there is a texture defined in the class, it will show the texture.
Changed Beer limits to show Game Over screen when the last mug of beer changes to an empty texture

Scenes now have a designated property to how many items and enemies there are in the scene to calculate player's score.

Fixed enemy rendering - When a player had killed all enemies, the game would crash due to using an int with a draw method